# MmioPatterns optimisation